[fix] #109 직원 삭제 서비스 코드 수정

+++ src/main/java/kyonggi/bookslyserver/domain/shop/service/EmployeeService.java
@@ -97,6 +97,10 @@ public class EmployeeService {
     public EmployeeDeleteResponseDto delete(Long id){
         Optional<Employee> employee = employeeRepository.findById(id);
 
+        if(!employee.isPresent()){
+            throw new EntityNotFoundException();
+        }
+
         Shop shop = employee.get().getShop();
         shop.getEmployees().remove(employee.get());
         employeeRepository.deleteById(id);
